requirements.txt
pyqt6>=6.2
ultralytics>=8.0.0
opencv-python>=4.5
pillow>=9.0
openpyxl>=3.0
requests>=2.28
numpy>=1.23

